requirements.txt
torch
diffusers>=0.24.0
transformers
accelerate
gradio
pillow
safetensors
